Enable short options
`getopt_long` accepts both a vector of long options and an option
string. For a long option, it would return the identifier provided via
the last field of the respective `struct option`. Normally, one would
make it return an equivalent short option.

However, for some reason the author choose arbitrary enum values for
each longoption and only use those for option identification during the
parse. At the same time, the option string passed to `getopt_long` (and
the fact that `getopt_long_only` wasn't used instead) does imply that
`pam-test` was supposed to accept a number of short options. What's
more, the option string does not account for parameters.

This commit introduces support for those short options, taking an
educated guess which short option a long option should translate to.